Move e2e deps set
